backend/requirements.txt
flask
flask_sqlalchemy
flask_migrate
flask_cors
pymysql
python-dotenv
psycopg2-binary
flask_jwt_extended
flask_mail
flask_bcrypt
flask_login